Make module to load environment configuration
